reverted server split (combined for 1?) and tried creating master node & quick float trunc into int

+++ workdir/Server.java
@@ -74,7 +74,7 @@ public class Server {
 
 		//designated master node, started required # of VM;s
 		if (id == 1) {
-			for (int i = 0; i < hardcoded[SL.getTime()]; i ++) {
+			for (int i = 0; i < hardcoded[(int) SL.getTime()]; i ++) {
 				SL.startVM();
 			}
 		}
